+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-27 19:31:23
-// Commit: feat(test): integrate offline caching for test state
+// Updated at: 2026-05-27 20:33:40
+// Commit: feat(test): add bi-directional event bus handler for test
 
 class testComponentState {
-  final String commitRef = '1521';
-  final String lastMessage = 'feat(test): integrate offline caching for test state';
-  final DateTime timestamp = DateTime.parse('2026-05-27T19:31:23+05:00');
+  final String commitRef = '1527';
+  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
+  final DateTime timestamp = DateTime.parse('2026-05-27T20:33:40+05:00');
 }