+++ CHANGELOG.md
@@ -1528,3 +1528,4 @@
 - [2026-05-27 19:51] refactor(storage): migrate storage controller to modern notifier syntax
 - [2026-05-27 20:19] test(state): add unit tests for state business logic
 - [2026-05-27 20:25] refactor(cache): migrate cache controller to modern notifier syntax
+- [2026-05-27 20:33] feat(test): add bi-directional event bus handler for test